v0.3 - Breaking Change (again)
The whole module has been simplified to use only the runtime version
as there's no way to reasonably maintain it via pre-defined set of
HTML, SVG, MathML, and who knows what else could come up as namespace.

All exports have been removed except for `/core` and `/ce` and by default
the `/core` export is what's being published as main entry.

It is possible that in the future the index would be one with all the features,
as opposite of just core, so for now stick with `/core` to be sure, or `/ce`,
as these are not planned to change in the future.